pyproject.toml
[project]
name = "profile-intelligence-stage1"
version = "0.1.0"
description = "Profile enrichment service using external APIs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.49",
    "uuid7>=0.1.0",
    "uvicorn[standard]>=0.44.0",
    "gunicorn>=21.2.0",
    "pytest>=9.0.3",
    "aiosqlite>=0.22.1",
    "pytest-asyncio>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]  # E=pycodestyle, F=pyflakes, I=isort

[tool.ruff.lint.isort]
known-first-party = ["app"]